fix(build): pass build
